chore(structure): refactored using ng-packagr
