+++ e2e/app.e2e-spec.ts
@@ -1,0 +1,14 @@
+import { AppPage } from './app.po';
+
+describe('ngx-localstorage App', () => {
+  let page: AppPage;
+
+  beforeEach(() => {
+    page = new AppPage();
+  });
+
+  it('should display welcome message', () => {
+    page.navigateTo();
+    expect(page.getParagraphText()).toEqual('Welcome to app!');
+  });
+});
