--- index.ts
@@ -1,4 +1,0 @@
-/**
- * Created by bohoffi on 03.04.2017.
- */
-export * from './src/ngx-localstorage.module';
