+++ package.json
@@ -1,13 +1,20 @@
 {
   "name": "ngx-localstorage",
-  "version": "1.2.0",
+  "version": "2.0.0",
   "description": "An Angular (>=5.0.0) wrapper for local storage access.",
+  "ngPackage": {
+    "lib": {
+      "entryFile": "public_api.ts"
+    }
+  },
   "scripts": {
-    "cleanup": "rimraf bundles",
-    "bundling": "rollup -c",
-    "minify": "uglifyjs bundles/ngx-localstorage.umd.js --screw-ie8 --compress --mangle --comments --output bundles/ngx-localstorage.umd.min.js",
-    "build": "npm run cleanup && ngc && npm run bundling && npm run minify",
-    "prepublish": "npm run build"
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build --prod",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
+    "packagr": "ng-packagr -p ng-package.json"
   },
   "repository": {
     "type": "git",
@@ -25,25 +32,40 @@
     "url": "https://github.com/bohoffi/ngx-localstorage/issues"
   },
   "homepage": "https://github.com/bohoffi/ngx-localstorage#readme",
-  "module": "index.js",
-  "main": "bundles/ngx-localstorage.umd.js",
-  "typings": "index.d.ts",
-  "dependencies": {},
-  "devDependencies": {
-    "@angular/common": "^5.2.1",
-    "@angular/compiler": "^5.2.1",
-    "@angular/compiler-cli": "^5.2.1",
-    "@angular/core": "^5.2.1",
-    "@types/core-js": "^0.9.45",
-    "rimraf": "^2.6.1",
-    "rollup": "^0.41.6",
+  "dependencies": {
+    "@angular/animations": "^5.2.0",
+    "@angular/common": "^5.2.0",
+    "@angular/compiler": "^5.2.0",
+    "@angular/core": "^5.2.0",
+    "@angular/forms": "^5.2.0",
+    "@angular/http": "^5.2.0",
+    "@angular/platform-browser": "^5.2.0",
+    "@angular/platform-browser-dynamic": "^5.2.0",
+    "@angular/router": "^5.2.0",
+    "core-js": "^2.4.1",
     "rxjs": "^5.5.6",
-    "typescript": "^2.5.3",
-    "uglify-js": "^3.3.8",
-    "zone.js": "^0.8.20"
+    "zone.js": "^0.8.19"
   },
-  "peerDependencies": {
-    "@angular/common": ">=5.0.0",
-    "@angular/core": ">=5.0.0"
+  "devDependencies": {
+    "@angular/cli": "1.6.5",
+    "@angular/compiler-cli": "^5.2.0",
+    "@angular/language-service": "^5.2.0",
+    "@types/jasmine": "~2.8.3",
+    "@types/jasminewd2": "~2.0.2",
+    "@types/node": "~6.0.60",
+    "codelyzer": "^4.0.1",
+    "jasmine-core": "~2.8.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~2.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-cli": "~1.0.1",
+    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-jasmine": "~1.1.0",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "ng-packagr": "^1.6.0",
+    "protractor": "~5.1.2",
+    "ts-node": "~4.1.0",
+    "tslint": "~5.9.1",
+    "typescript": "~2.5.3"
   }
 }
