+++ src/app/modules/demo/decorator/decorator.component.html
@@ -1,0 +1,7 @@
+<code>
+  <span>@ngxLocalStorage(&#123;key: 'cbox1', prefix: 'demo'&#125;)</span><br>
+  <span>boundCheckBox: boolean;</span>
+</code>
+<p>
+  Decorator: {{boundCheckBox}}
+</p>
