+++ src/app/modules/demo/directive/directive.component.html
@@ -1,0 +1,8 @@
+<code>
+
+</code>
+
+<p>
+  <input type="checkbox" id="cbox1" ngxLocalStorage lsPrefix="demo" lsInitFromStorage="true" lsEvent="change" [lsValuePath]="['checked']">
+  <label for="cbox1">Checkbox</label>
+</p>
