+++ src/app/modules/demo/directive/directive.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-directive',
+  templateUrl: './directive.component.html',
+  styleUrls: ['./directive.component.less']
+})
+export class DirectiveComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
