+++ src/app/modules/ngx-localstorage/lib/directives/index.ts
