+++ src/app/modules/ngx-localstorage/lib/index.ts
